Rename `_repository.py` to `repository.py`
